# openapi.yaml
openapi: 3.0.3
info:
  title: Otter PPT API
  version: 0.1.0
  description: Generate editable PowerPoint files using configured models, external tool calls, or presentation JSON. AI credentials are optional for execute/build.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/execute:
    post:
      operationId: executeToolCalls
      summary: Apply externally generated tool calls without invoking an AI model (stateless)
      description: |
        Stateless — pass the full `presentation` and the updated state is returned.
        For iterative editing (and server-side undo/redo) prefer the stateful
        session endpoints: POST /api/v1/session, then POST /api/v1/session/{id}/execute.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExecuteRequest' }
      responses:
        '200':
          description: Updated presentation and tool results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecuteResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /api/v1/render:
    post:
      operationId: renderPresentation
      summary: Render presentation JSON to slide images (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Presentation' }
      responses:
        '200':
          description: Rendered slide images
          content:
            application/json:
              schema:
                type: object
                properties:
                  backend: { type: string }
                  slide_count: { type: integer }
                  slides:
                    type: array
                    items:
                      type: object
                      properties:
                        slide_num: { type: integer }
                        width: { type: integer }
                        height: { type: integer }
                        image_base64: { type: string }
                        description: { type: string }
        '400': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/session:
    post:
      operationId: createSession
      summary: Create a stateful editing session (undo/redo without reshipping state)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                presentation: { $ref: '#/components/schemas/Presentation' }
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                required: [session_id]
                properties:
                  session_id: { type: string }
                  expires_in: { type: string, example: 30m0s }
                  slide_count: { type: integer }
                  usage: { type: object, additionalProperties: true }
...
  /api/v1/session/{id}/execute:
    post:
      operationId: executeSessionToolCalls
      summary: Run tool calls against a server-side session (no state round-trip)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [calls]
              properties:
                calls:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/ToolCall' }
      responses:
        '200':
          description: Tool results and updated presentation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecuteResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /api/v1/session/{id}/render:
    post:
      operationId: renderSession
      summary: Render the session's current state to slide images
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Rendered slide images (same shape as /api/v1/render) }
        '404': { $ref: '#/components/responses/Error' }
...
  /api/v1/session/{id}/undo:
    post:
      operationId: undoSession
      summary: Undo the last mutating tool call
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Undo result }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
components:
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ToolCall:
      type: object
      required: [name]
      properties:
        name: { type: string }
        arguments: { type: object, additionalProperties: true }
    ExecuteRequest:
      type: object
      required: [calls]
      properties:
        presentation: { $ref: '#/components/schemas/Presentation' }
        calls:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ToolCall' }
    ExecuteResponse:
      type: object
      required: [presentation, results]
      properties:
        presentation: { $ref: '#/components/schemas/Presentation' }
        results:
          type: array
          items: { type: object, additionalProperties: true }
...
    Presentation:
      type: object
      required: [title, theme, slides]
      properties:
        title: { type: string }
        theme: { $ref: '#/components/schemas/Theme' }
        slides: { type: array, items: { $ref: '#/components/schemas/Slide' } }
        slide_width: { type: number, default: 13.333 }
        slide_height: { type: number, default: 7.5 }
    Theme:
      type: object
      required: [primary_color, secondary_color, accent_color, background_color, text_color, title_font, body_font]
      properties:
        name: { type: string }
        primary_color: { type: string }
        secondary_color: { type: string }
        accent_color: { type: string }
        background_color: { type: string }
        text_color: { type: string }
        title_font: { type: string }
        body_font: { type: string }
    Slide:
      type: object
      required: [id, layout, elements]
      properties:
        id: { type: string }
        layout: { type: string }
        background: { type: object, additionalProperties: true }
        transition: { type: object, additionalProperties: true }
        notes: { type: string }
        elements:
          type: array
          items: { type: object, additionalProperties: true }